Move to double type and set target to cero on creation

+++ src/Pid.h
@@ -8,22 +8,22 @@
 class Pid {
   public:
 	Pid();
-	void setTarget(float target);
-	void setKp(float Kp);
-	void setKi(float Ki);
-	void setKd(float Kd);
-	void setMaxWindup(float max_windup);
-	void setAlpha(float alpha);
-	float update(float measure,float dt);
+	void setTarget(double target);
+	void setKp(double Kp);
+	void setKi(double Ki);
+	void setKd(double Kd);
+	void setMaxWindup(double max_windup);
+	void setAlpha(double alpha);
+	double update(double measure,double dt);
 	void reset();
   private:
-  	float Kp_; 	
-  	float Ki_;
-  	float Kd_;
-  	float max_windup_;
-  	float alpha_;
-  	float target_;
-  	float error_sum_;
-  	float error_last_;
+  	double Kp_; 	
+  	double Ki_;
+  	double Kd_;
+  	double max_windup_;
+  	double alpha_;
+  	double target_;
+  	double error_sum_;
+  	double error_last_;
 };
 #endif
